lab4_nosql/requirements.txt
Django==3.1.12
djangorestframework==3.14.0
python-decouple==3.8
djongo==1.3.7
setuptools==71.1.0
pillow==10.4.0
python-magic==0.4.27
django-resized==1.0.3